requirements.txt
altgraph==0.17.2
aniso8601==9.0.1
asgiref==3.5.2
astroid==2.11.7
async-generator==1.10
attrs==21.4.0
auto-py-to-exe==2.20.1
autopep8==1.6.0
beautifulsoup4==4.11.1
bottle==0.12.23
bottle-websocket==0.2.9
bs4==0.0.1
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
constants==0.6.0
cryptography==36.0.2
cssselect2==0.6.0
cycler==0.11.0
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
Django==4.0.6
django-ckeditor==6.5.1
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==2.0.0
django-phonenumber-field==7.0.0
django-widget-tweaks==1.4.12
docutils==0.19
Eel==0.14.0
filelock==3.7.1
flake8==4.0.1
Flask==2.2.2
fonttools==4.34.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
gitdb==4.0.9
GitPython==3.1.27
graphene==3.1.1
graphene-django==3.0.0b7
graphql-core==3.2.1
graphql-relay==3.2.0
greenlet==1.1.3
h11==0.13.0
iconify==0.0.103
idna==2.10
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kaki==0.1.5
kids.cache==0.0.7
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
kivy-designer==0.1.dev0
Kivy-examples==2.1.0
Kivy-Garden==0.1.5
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
kivymd-extensions.sweetalert==0.1.5
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
libsass==0.21.0
lk-lambdex==2.3.0
lk-logger==5.1.6
lk-qtquick-scaffold==1.3.0
lk-utils==2.3.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
mock==4.0.3
monotonic==1.6
MouseInfo==0.1.3
numpy==1.23.1
outcome==1.1.0
packaging==21.3
parso==0.8.3
pefile==2022.5.30
phonenumbers==8.12.54
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
plyer==2.0.0
promise==2.3
PyAutoGUI==0.9.53
pycairo==1.21.0
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.12.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.9
pylint==2.14.5
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.3.0
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
PyRect==0.2.0
PyScreeze==0.1.28
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
pytoolconfig==1.2.2
pytweening==1.0.4
pyuic5-tool==0.0.1
pywhatkit==5.4
pywin32==304
pywin32-ctypes==0.2.0
QT-PyQt-PySide-Custom-Widgets==0.5.7
QtPy==2.1.0
qtsass==0.3.0
requests==2.23.0
rich==12.5.1
rope==1.3.0
selenium==4.1.3
shiboken2==5.15.2.1
shiboken6==6.3.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
ujson==5.4.0
urllib3==1.25.11
virtualenv==20.16.2
watchdog==2.1.9
webencodings==0.5.1
Werkzeug==2.2.2
whatthepatch==1.0.2
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.1.0
xcffib==0.11.1
yapf==0.32.0
zope.event==4.5.0
zope.interface==5.4.0
